Binds a machine to AD and migrates the users data
This script was written to be used in Casper Suite 9, and has been
tested with OS X 10.8-10.10.  This script will bind the Mac to AD and
migrate the users account from a local account to a network managed
account.  This script assumes the local user name and AD user name are
the same.